search places by keyword works for now
